bot/requirements.txt
aiogram==3.27.0
aiosqlite==0.18.0
python-dotenv==1.0.1
